pyproject.toml
[tool.poetry]
name = "opentelemetry-ml"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opentelemetry-api = "^0.12b0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.4.2"
scikit-learn = "^0.23.2"
opentelemetry-sdk = "^0.12b0"
opentelemetry-exporter-jaeger = "^0.12b0"
fastapi = "^0.61.1"
opentelemetry-instrumentation-fastapi = "^0.12b0"
uvicorn = "^0.11.8"
opentelemetry-test = { path = "../opentelemetry-python/tests/util" }
torch = "^1.6.0"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.circleci
    | \.git
    | \.github
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
force_single_line = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip = [".venv"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


